Create sidebar section with tasks and categories
created three main divs inside the first section
which is dedicated to like the logo of the website
the all tasks section completed tasks statistics
section and a categories section that includes things
like home category, fitness one, work etc.
a add task button with a plus icon to indicate that.