Create search bar and dark theme toggle section
made a section with search bar and a radio buttons
toggle that between dark theme and lighttheme.
the search bar and the ui toggle buttons all of them
have icons telling or mimicking the scenaria
which means all of the icons images i have added to
them in two versions lighttheme and dark theme.