Add font family and link it to the html
added the font family inter to be my main font and
linked it to html. i solved alittle problem which
was a closing that didn't had an opening one. it was
not a major deal but corrected it. also prettier
formatter which i used messed things up like so if
you see a lot of like red and green pluses, it doesn't
mean i changed a ton of things. it is just the prettier.