Style the logo and name section of sidebar
added classes to the div that houses the logo of
website and the name of it, also the logo text.
the main div .logo-div i made it flex and aligned
it to the center so that the logo and the text aligns
to center of the div vertically. resized the default
logo image to become small enough for content. the
text logo name size is 1.5rem with a fontweight of
700 and margin left of 1rem so it has a nice gap
between the logo image and the text