Add icons to tasks div and style them to match the UI
add icons to these paragraphs(All tasks, completed,
today, statistics) to make them appealing and good.

also the the paragraphs, i made the display of flex
and font weight of 500.

styled the svg icons like the following. first, i
made the height and width 1.25rem, and stroke color
of #525252. for the stroke width. i made the changes
inside the svgs which was the easiest choice for me now.

also i gave the svg icons a margin rigt of 1rem so
it has enough room between the icon and the text.